+++ routes/controls.ts
@@ -59,7 +59,7 @@ function defineControls(categories: Category[]) {
       ...r,
       id: String(r.id),
       category: cat.name,
-      startTime: r.starttime || '',
+      startTime: r.startTime || '',
       yearOfBirth: r.yearOfBirth?.toString(),
       splits: r.splits || []
     }));
